Create seperate Formatter class to handle the formatting in TextUi class.
